# openapi.yaml
openapi: 3.1.0
info:
  title: SparkyPod API
  description: |
    REST API for managing podcast feeds with automatic ad detection and removal.


    This API allows you to:
    - Manage podcast feed subscriptions
    - View and manage processed episodes
    - Configure ad detection settings including AI model selection
    - Monitor system status and trigger cleanup operations
    - Manage cross-episode ad patterns with network and podcast scope
    - Submit corrections to improve ad detection accuracy
  version: 2.16.0
  contact:
    name: SparkyPod
  license:
    name: MIT
servers:
  # Same-origin relative URL. Swagger UI served from /api/v1/docs uses
  # whatever host the browser is on, so "Try it out" always hits the
  # correct server: localhost in dev, the production domain in prod,
  # a Tailscale/Cloudflare URL on a tunnel, etc. No hardcoded host.
  - url: /api/v1
    description: Same-origin (current server)
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: '{baseUrl}/api/v1'
    description: Custom server
    variables:
      baseUrl:
        default: http://localhost:8000
        description: Base URL of the podcast server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login with password
      description: Authenticate with the configured password. Rate limited to 5 attempts per minute.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  description: The application password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/password:
    put:
      tags:
        - Auth
      summary: Set or change password
      description: |
        Set or change the application password. Rate limited to 3 attempts per hour.

        - If no password is set, creates a new password
        - If password is set, requires current password for verification
        - Set newPassword to empty string to remove password protection
        - Minimum password length: 8 characters
      operationId: setPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newPassword
              properties:
                currentPassword:
                  type: string
                  description: Current password (required if password is already set)
                newPassword:
                  type: string
                  description: New password (min 8 chars, or empty to remove protection)
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  passwordSet:
                    type: boolean
        '400':
          description: Invalid request (password too short)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /feeds:
    get:
      tags:
        - Feeds
      summary: List all feeds
      description: Returns a list of all configured podcast feeds with metadata
      operationId: listFeeds
      responses:
        '200':
          description: List of feeds
          content:
            application/json:
              schema:
                type: object
                properties:
                  feeds:
                    type: array
                    items:
                      $ref: '#/components/schemas/Feed'
    post:
      tags:
        - Feeds
      summary: Add a new feed
      description: Add a new podcast feed by providing the source RSS URL
      operationId: addFeed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sourceUrl
              properties:
                sourceUrl:
                  type: string
                  format: uri
                  description: URL of the source RSS feed
                  example: https://feeds.example.com/podcast.xml
                slug:
                  type: string
                  description: Optional custom slug (auto-generated if not provided)
                  example: my-podcast
                maxEpisodes:
                  type: integer
                  minimum: 10
                  maximum: 500
                  description: Per-feed cap. Omit or set null to fall back to the global maxFeedEpisodes setting (default 300).
                onlyExposeProcessedEpisodes:
                  type:
                    - boolean
                    - 'null'
                  description: Per-feed override. true hides non-processed episodes; false shows everything; null falls back to the global onlyExposeProcessedDefault setting.
      responses:
        '201':
          description: Feed created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  slug:
                    type: string
                  sourceUrl:
                    type: string
                  feedUrl:
                    type: string
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Feed with slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /feeds/{slug}:
    get:
      tags:
        - Feeds
      summary: Get a single feed
      description: Returns details for a specific podcast feed
      operationId: getFeed
      parameters:
        - $ref: '#/components/parameters/slug'
      responses:
        '200':
          description: Feed details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feed'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Feeds
      summary: Delete a feed
      description: Delete a podcast feed and all associated data
      operationId: deleteFeed
      parameters:
        - $ref: '#/components/parameters/slug'
      responses:
        '200':
          description: Feed deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  slug:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - Feeds
      summary: Update feed settings
      description: Update podcast feed settings like network override, DAI platform, and processing overrides
      operationId: updateFeed
      parameters:
        - $ref: '#/components/parameters/slug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                networkIdOverride:
                  type:
                    - string
                    - 'null'
                  description: >-
                    Network to assign this feed to: a known network id, or a
                    custom name that links feeds sharing the same value (null
                    uses the auto-detected network). The effective network is
                    the override when set, else the auto-detected one.
                daiPlatform:
                  type: string
                  description: DAI platform identifier (e.g., megaphone, acast)
                autoProcessOverride:
                  type:
                    - boolean
                    - 'null'
                  description: Override auto-process setting (null uses global)
                languageOverride:
                  type:
                    - string
                    - 'null'
                  description: Per-feed Whisper transcription language. 'auto' pins auto-detect for this feed; a language code (e.g. 'en', 'de') forces that language; null falls back to the global whisperLanguage setting.
                titleOverride:
                  type:
                    - string
                    - 'null'
                  maxLength: 500
                  description: Per-feed display title shown to subscribers in the generated RSS feed (issue 375). Survives RSS refreshes. Null or empty clears it and falls back to the source feed title.
                detectionMode:
                  type:
                    - string
                    - 'null'
                  enum:
                    - blacklist
                    - keep_content
                    - null
                  description: Per-feed ad-detection mode. 'keep_content' (experimental) has the LLM label show content and removes everything else; null or 'blacklist' is the default, which labels and removes ads.
                maxEpisodes:
                  type:
                    - integer
                    - 'null'
                  minimum: 10
                  maximum: 500
                  description: Per-feed cap. Set null to fall back to the global maxFeedEpisodes setting.
                onlyExposeProcessedEpisodes:
                  type:
                    - boolean
                    - 'null'
                  description: Per-feed override. true hides non-processed episodes; false shows everything; null falls back to the global onlyExposeProcessedDefault setting.
      responses:
        '200':
          description: Feed updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feed'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /system/backup:
    get:
      tags:
        - System
      summary: Download database backup
      description: |
        Returns a copy of the SQLite database file for backup purposes. Rate limited to 6/hour.
      operationId: downloadBackup
      responses:
        '200':
          description: Database backup file download
          headers:
            Content-Disposition:
              schema:
                type: string
              description: 'Attachment filename, e.g. attachment; filename="sparkypod-backup-2026-03-12.db"'
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        Flask session cookie issued by ``POST /auth/login``. Every mutating
        endpoint additionally requires a matching ``X-CSRF-Token`` header
        (double-submit). Public endpoints (``/auth/login``, ``/auth/logout``,
        ``/auth/status``, ``/health``, ``/health/live``) declare ``security: []``.
  parameters:
    slug:
      name: slug
      in: path
      required: true
      description: Podcast feed slug identifier
      schema:
        type: string
        example: my-podcast
...
  schemas:
...
    Feed:
      type: object
      properties:
        slug:
          type: string
          description: URL-safe identifier for the feed
        title:
          type: string
          description: Podcast title from RSS feed
        sourceUrl:
          type: string
          format: uri
          description: Original RSS feed URL
        feedUrl:
          type: string
          format: uri
          description: Modified RSS feed URL served by this server
        artworkUrl:
          type:
            - string
            - 'null'
          description: URL to cached artwork image
        episodeCount:
          type: integer
          description: Total number of episodes
        processedCount:
          type: integer
          description: Number of processed episodes
        lastRefreshed:
          type:
            - string
            - 'null'
          format: date-time
          description: Last RSS refresh timestamp
        createdAt:
          type: string
          format: date-time
        autoProcessOverride:
          type:
            - boolean
            - 'null'
          description: Override auto-process setting for this feed (null uses global setting)
        languageOverride:
          type:
            - string
            - 'null'
          description: Per-feed Whisper transcription language ('auto' or a language code like 'en'/'de'); null uses the global whisperLanguage setting.
        titleOverride:
          type:
            - string
            - 'null'
          description: Per-feed display title shown to subscribers in the generated RSS feed (issue 375); null falls back to the source feed title.
        detectionMode:
          type:
            - string
            - 'null'
          enum:
            - blacklist
            - keep_content
            - null
          description: Per-feed ad-detection mode. 'keep_content' (experimental) labels show content and removes everything else; null or 'blacklist' is the default ad-removal behavior.
        maxEpisodes:
          type:
            - integer
            - 'null'
          minimum: 10
          maximum: 500
          description: Per-feed cap on episodes served via modified RSS. Null means use the global maxFeedEpisodes setting (default 300).
        onlyExposeProcessedEpisodes:
          type:
            - boolean
            - 'null'
          description: Per-feed override for hiding upstream episodes whose status is not 'processed' from the served RSS feed. Null means use the global onlyExposeProcessedDefault setting. true/false override the global for this feed (issue #181).
...
    Error:
      type: object
      properties:
        error:
          type: string
        status:
          type: integer
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...